Test_Complete
10개 동시 실행시 1개만 적용되는 사항은 테스트 실패 / 예약 중복시 IllegalStateException 발생 확인

+++ src/main/java/com/example/getstudyroom/dto/RoomDto.java
@@ -19,6 +19,6 @@ public class RoomDto {
         private String location;
 
         @Positive(message = "수용인원은 1 이상이어야 합니다.")
-        private Long capacity;
+        private int capacity;
     }
 }
